Soil Detection/requirements.txt
Flask==2.0.2
paho-mqtt==1.6.1
requests==2.26.0
